2016-11-23a Add user defined transit time between Earth and Mars (tentatively at 60 sols). Fix the use of average transit time. Add displaying the launch date of a resupply mission. Pad the orbit (Martian year) with zero so as to always display 4 digits.
# NEW

1. Add launchDateValueLabel in ResupplyDetailPanel. 

2. Add getAverageTransitTime() in SimulationConfig. 

3. Add average-transit-time in SimulationConfig and simulation.xml and simulation.dtd. 

4. Add MAX_NUM_SOLS_PLANNED in ResupplyUtil. 

5. Add updateSolsCB() in ResupplyMissionEditingPanel. 

6. Append padding zeros to orbit in getDateString() in MarsClock.
e.g. "0013-Adir-05" instead of "13-Adir-05"
This way, one can tell 0013 is the year. Much like 2016-Jan-05


# CHANGES

1. Modify AVG_TRANSIT_TIME to load user defined value by calling  SimulationConfig.instance().getAverageTransitTime(). 

2. Switch autosave_minute from long to int.


# FIX

1. Allow user to choose only value larger than average-transit-time in solsCB in ResupplyMissionEditingPanel.

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/Simulation.java
@@ -103,7 +103,7 @@ implements ClockListener, Serializable {
             File.separator +
             Msg.getString("Simulation.defaultDir.autosave"); //$NON-NLS-1$
 
-	public static long autosave_minute;// = 15;
+	public static int autosave_minute;// = 15;
 
 	// Categories of loading and saving simulation
 	public static final int OTHER = 0; // load other file
